+++ src/components/molecules/PlayInfo/PlayInfo.tsx
@@ -82,6 +82,7 @@ const useStyles = createUseStyles((theme) => ({
 // TODO: add long name marquee
 // TODO: add Artwork magnifier UI
 // TODO: add Artwork link to currently playing main content
+// TODO: add favorite track heart button
 
 const PlayInfo = ({ className, songDetails }) => {
   const theme = useTheme()