+++ src/components/organisms/SongList/SongList.tsx
@@ -8,7 +8,7 @@ import moment from 'moment'
 
 import { Icon } from '../../atoms'
 import { 
-  fetchPlaylistSongsPending,
+  // fetchPlaylistSongsPending,
   fetchSongs,
   fetchSongsError,
   fetchSongsPending