+++ src/containers/MainView/index.ts
@@ -1,10 +1,3 @@
 import MainView from './MainView'
-import { connect } from 'react-redux'
 
-const mapStateToProps = (state) => {
-  return {
-    headerTitle: state.uiReducer.title
-  }
-}
-
-export default connect(mapStateToProps)(MainView)
+export default MainView